mana calculations and probabilities
this large push includes progress towards calculating the probability of drawing lands that produce a certain combination of mana. this works by first determining every possible combination of lands that can produce the desired mana, and then uses a monte carlo simulation to determine probabilities. the monte carlo simulation was checked for accuracy against the hypergeometric functions and appears correct (but this can only test 1 combination).

I tried a more fancy method called the inclusion-exclusion principle, but i think this will be too computationally intense. monte carlo appears like it'll work well enough, but will have to decide on the iteration count (and if user can set the interation count).